+++ nengo/neurons.py
@@ -57,18 +57,35 @@ class NeuronType(FrozenObject):
     def current(self, x, gain, bias):
         """Compute current injected in each neuron given input, gain and bias.
 
+        Note that ``x`` is assumed to be already projected onto the encoders
+        associated with the neurons and normalized to radius 1, so the maximum
+        expected current for a neuron occurs when input for that neuron is 1.
+
         Parameters
         ----------
-        x : (n_neurons,) array_like
-            Vector-space input.
+        x : (n_samples,) or (n_samples, n_neurons) array_like
+            Scalar inputs for which to calculate current.
         gain : (n_neurons,) array_like
             Gains associated with each neuron.
         bias : (n_neurons,) array_like
             Bias current associated with each neuron.
+
+        Returns
+        -------
+        current : (n_samples, n_neurons)
+            Current to be injected in each neuron.
         """
         x = np.array(x, dtype=float, copy=False, ndmin=1)
         gain = np.array(gain, dtype=float, copy=False, ndmin=1)
         bias = np.array(bias, dtype=float, copy=False, ndmin=1)
+
+        if x.ndim == 1:
+            x = x[:, np.newaxis]
+        elif x.ndim >= 3 or x.shape[1] != gain.shape[0]:
+            raise ValidationError("Expected shape (%d, %d); got %s."
+                                  % (x.shape[0], gain.shape[0], x.shape),
+                                  attr="x", obj=self)
+
         return gain * x + bias
 
     def gain_bias(self, max_rates, intercepts):
@@ -102,9 +119,8 @@ class NeuronType(FrozenObject):
         max_rate = max_rates.max()
 
         # Start with dummy gain and bias so x == J in rate calculation
-        gain = np.ones(J_steps)
-        bias = np.zeros(J_steps)
-        rate = np.zeros(J_steps)
+        gain = np.ones(1)
+        bias = np.zeros(1)
 
         # Find range of J that will achieve max rates (assume monotonic)
         J_threshold = None
@@ -128,11 +144,12 @@ class NeuronType(FrozenObject):
                 raise RuntimeError("Could not find max current")
 
         J = np.linspace(J_threshold, J_max, J_steps)
-        rate = self.rates(J, gain, bias)
+        rate = self.rates(J, gain, bias).squeeze(axis=1)
 
         gain = np.zeros_like(max_rates)
         bias = np.zeros_like(max_rates)
         J_tops = np.interp(max_rates, rate, J)
+
         gain[:] = (J_threshold - J_tops) / (intercepts - 1)
         bias[:] = J_tops - gain
         return gain, bias
@@ -158,28 +175,31 @@ class NeuronType(FrozenObject):
             X-intercepts of neurons.
         """
 
-        max_rates = self.rates(np.ones_like(gain), gain, bias)
+        max_rates = self.rates(1, gain, bias).squeeze(axis=0)
 
         x_range = np.linspace(-1, 1, 101)
-        rates = np.asarray([self.rates(np.ones_like(gain) * x, gain, bias)
-                            for x in x_range])
+        rates = self.rates(x_range, gain, bias)
         last_zeros = np.maximum(np.argmax(rates > 0, axis=0) - 1, 0)
         intercepts = x_range[last_zeros]
 
         return max_rates, intercepts
 
     def rates(self, x, gain, bias):
-        """Compute firing rates (in Hz) for given vector input, ``x``.
+        """Compute firing rates (in Hz) for given input ``x``.
 
         This default implementation takes the naive approach of running the
         step function for a second. This should suffice for most rate-based
         neuron types; for spiking neurons it will likely fail (those models
         should override this function).
 
+        Note that ``x`` is assumed to be already projected onto the encoders
+        associated with the neurons and normalized to radius 1, so the maximum
+        expected rate for a neuron occurs when input for that neuron is 1.
+
         Parameters
         ----------
-        x : (n_neurons,) array_like
-            Vector-space input.
+        x : (n_samples,) or (n_samples, n_neurons) array_like
+            Scalar inputs for which to calculate rates.
         gain : (n_neurons,) array_like
             Gains associated with each neuron.
         bias : (n_neurons,) array_like
@@ -187,7 +207,7 @@ class NeuronType(FrozenObject):
 
         Returns
         -------
-        rates : (n_neurons,) ndarray
+        rates : (n_samples, n_neurons) ndarray
             The firing rates at each given value of ``x``.
         """
         J = self.current(x, gain, bias)