+++ CHANGES.rst
@@ -92,6 +92,9 @@ Release History
   (`#1521 <https://github.com/nengo/nengo/pull/1521>`_)
 - Switched to nengo-bones templating system for TravisCI config/scripts.
   (`#1514 <https://github.com/nengo/nengo/pull/1514>`_)
+- The ``NeuronType.current`` and ``NeuronType.rates`` methods now document
+  the supported shapes of parameters and return values.
+  (`#1437 <https://github.com/nengo/nengo/pull/1437>`__)
 
 **Deprecated**